add some geometry and mesh do some update

+++ meshGenerate/src/Surface.h
@@ -14,13 +14,13 @@ class Surface
 private:
 	unsigned int m_Id;
 	SurfaceType m_type;
-	long long m_PointNumber;
+	size_t m_PointNumber;
 
-	std::vector<long long> m_PointIds;
+	std::vector<size_t> m_PointIds;
 
 public:
 	Surface(unsigned int id, SurfaceType type, unsigned int *PointId);
 
-	inline long long Get_PointNumber() const { return m_PointNumber; }
-	inline std::vector<long long> Get_PointIds() const { return m_PointIds; }
+	inline size_t Get_PointNumber() const { return m_PointNumber; }
+	inline std::vector<size_t> Get_PointIds() const { return m_PointIds; }
 };
